@@ -4,139 +4,81 @@ namespace Laralabs\GetAddress\Responses;
 
 class Address
 {
-    /**
-     * Sort returned addresses numerically.
-     *
-     * @var bool
-     */
-    const SORT_NUMERICALLY = true;
+    protected array $address = [];
 
-    /**
-     * Dont perform any specific sort on the returned addresses.
-     *
-     * @var bool
-     */
-    const NO_SORT = false;
-
-    /**
-     * Address string.
-     *
-     * @var array
-     */
-    protected $address = [];
-
-    /**
-     * Constructor.
-     *
-     * @param string $address
-     *
-     * @return void
-     */
-    public function __construct($address)
+    public function __construct(string|array $address)
     {
-        $this->address = explode(',', $address);
+        $this->address = is_string($address) ? explode(',', $address) : $address;
     }
 
-    /**
-     * Get Line 1.
-     *
-     * @return string
-     */
     public function getLine1(): string
     {
-        return $this->address[0];
+        return $this->address[0] ?? $this->address['line_1'];
     }
 
-    /**
-     * Get Line 2.
-     *
-     * @return string
-     */
     public function getLine2(): string
     {
-        return $this->address[1];
+        return $this->address[1] ?? $this->address['line_2'];
     }
 
-    /**
-     * Get Line 3.
-     *
-     * @return string
-     */
     public function getLine3(): string
     {
-        return $this->address[2];
+        return $this->address[2] ?? $this->address['line_3'];
     }
 
-    /**
-     * Get Line 4.
-     *
-     * @return string
-     */
     public function getLine4(): string
     {
-        return $this->address[3];
+        return $this->address[3] ?? $this->address['line_4'];
     }
 
-    /**
-     * Get Line.
-     *
-     * @param int $line
-     *
-     * @return string
-     */
-    public function getLine($line): string
+    public function getLine(int $line): ?string
     {
-        return $this->address[$line - 1];
+        if ($line > 4) {
+            return null;
+        }
+
+        return $this->address[$line - 1] ?? null;
     }
 
-    /**
-     * Get Locality.
-     *
-     * @return string
-     */
     public function getLocality(): string
     {
-        return $this->address[4];
+        return $this->address[4] ?? $this->address['locality'];
     }
 
-    /**
-     * Get Town.
-     *
-     * @return string
-     */
     public function getTown(): string
     {
-        return $this->address[5];
+        return $this->address[5] ?? $this->address['town_or_city'];
     }
 
-    /**
-     * Get City.
-     *
-     * @return string
-     *
-     * @see Address:getTown()
-     */
     public function getCity(): string
     {
-        return $this->address[5];
+        return $this->address[5] ?? $this->address['town_or_city'];
     }
 
-    /**
-     * County.
-     *
-     * @return string
-     */
     public function getCounty(): string
     {
-        return $this->address[6];
+        return $this->address[6] ?? $this->address['county'];
+    }
+
+    public function getDistrict(): ?string
+    {
+        return $this->address[7] ?? $this->address['district'] ?? null;
+    }
+
+    public function getCountry(): ?string
+    {
+        return $this->address[8] ?? $this->address['country'] ?? null;
+    }
+
+    public function isResidential(): bool
+    {
+        return $this->address[9] ?? $this->address['residential'];
     }
 
     /**
      * Return a formatted array for the address.
      *
      * @param array $keys Override default key names
-     *
-     * @return array
      */
     public function toArray(array $keys = []): array
     {
@@ -152,37 +94,24 @@ class Address
      * Returns a string based on the address.
      *
      * @param bool $removeEmptyElements Prevents strings having conjoining commas
-     *
-     * @return string
      */
-    public function toString($removeEmptyElements = false): string
+    public function toString(bool $removeEmptyElements = false): string
     {
-        if (!$removeEmptyElements) {
+        if ($removeEmptyElements === false) {
             return implode(',', $this->address);
         }
 
-        return implode(',', array_filter($this->address, function ($value) {
-            return trim($value) !== '';
-        }));
+        return implode(',', array_filter(
+            $this->address,
+            static fn (?string $value): bool => filled($value)
+        ));
     }
 
-    /**
-     * Compare two addresses to see if they are equal.
-     *
-     * @param \Laralabs\GetAddress\Responses\Address $address Address to compare
-     *
-     * @return bool
-     */
     public function sameAs(self $address): bool
     {
         return !array_diff($this->address, $address->toArray());
     }
 
-    /**
-     * Convert the address to a comma separated string.
-     *
-     * @return string
-     */
     public function __toString(): string
     {
         return $this->toString();
