@@ -2,12 +2,9 @@
 
 namespace Laralabs\GetAddress\Exceptions;
 
-class ForbiddenException extends \Exception
+use Exception;
+
+class Forbidden extends Exception
 {
-    /**
-     * Exception Message.
-     *
-     * @var string
-     */
     protected $message = 'Your API key is not valid for this request.';
 }
