@@ -2,12 +2,9 @@
 
 namespace Laralabs\GetAddress\Exceptions;
 
-class InvalidPostcodeException extends \Exception
+use Exception;
+
+class InvalidPostcode extends Exception
 {
-    /**
-     * Exception Message.
-     *
-     * @var string
-     */
     protected $message = 'The postcode you provided was not a valid format.';
 }
