@@ -2,12 +2,9 @@
 
 namespace Laralabs\GetAddress\Exceptions;
 
-class ServerException extends \Exception
+use Exception;
+
+class ServerError extends Exception
 {
-    /**
-     * Exception Message.
-     *
-     * @var string
-     */
     protected $message = 'getAddress.io is currently having issues.';
 }
