@@ -2,12 +2,9 @@
 
 namespace Laralabs\GetAddress\Exceptions;
 
-class TooManyRequestsException extends \Exception
+use Exception;
+
+class TooManyRequests extends Exception
 {
-    /**
-     * Exception Message.
-     *
-     * @var string
-     */
     protected $message = 'You have made too many requests for this key.';
 }
