@@ -2,18 +2,18 @@
 
 namespace Laralabs\GetAddress\Models;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Laralabs\GetAddress\Factories\CachedAddressFactory;
 
 class CachedAddress extends Model
 {
-    /**
-     * @var string
-     */
+    use HasFactory;
+
+    /** @var string */
     protected $table = 'getaddress_cache';
 
-    /**
-     * @var array
-     */
+    /** @var array */
     protected $fillable = [
         'line_1',
         'line_2',
@@ -35,10 +35,7 @@ class CachedAddress extends Model
         'expanded_result',
     ];
 
-    /**
-     * @var array
-     */
-    public static $expandedFields = [
+    public static array $expandedFields = [
         'thoroughfare',
         'building_name',
         'sub_building_name',
@@ -55,10 +52,7 @@ class CachedAddress extends Model
         'country',
     ];
 
-    /**
-     * @var array
-     */
-    public static $fields = [
+    public static array $fields = [
         'line_1',
         'line_2',
         'line_3',
@@ -68,9 +62,6 @@ class CachedAddress extends Model
         'county',
     ];
 
-    /**
-     * @return string
-     */
     public function getFormattedStringAttribute(): string
     {
         return $this->toString(true);
@@ -80,15 +71,18 @@ class CachedAddress extends Model
      * Returns a string based on the address.
      *
      * @param bool $removeEmptyElements Prevents strings having conjoining commas
-     *
-     * @return string
      */
-    public function toString($removeEmptyElements = false): string
+    public function toString(bool $removeEmptyElements = false): string
     {
-        if (!$removeEmptyElements) {
+        if ($removeEmptyElements === false) {
             return implode(',', $this->only(self::$fields));
         }
 
         return implode(', ', array_filter($this->only(self::$fields)));
+    }
+
+    protected static function newFactory(): CachedAddressFactory
+    {
+        return new CachedAddressFactory;
     }
 }
