@@ -2,186 +2,93 @@
 
 namespace Laralabs\GetAddress\Responses;
 
-class ExpandedAddress
-{
-    /**
-     * Address string.
-     *
-     * @var array
-     */
-    protected $address = [];
+use Illuminate\Contracts\Support\Arrayable;
 
-    /**
-     * Constructor.
-     *
-     * @param string $address
-     *
-     * @return void
-     */
-    public function __construct($address)
+class ExpandedAddress implements Arrayable
+{
+    public function __construct(protected array $address = [])
     {
-        $this->address = $address;
     }
 
-    /**
-     * Get Thoroughfare.
-     *
-     * @return string
-     */
     public function getThoroughfare(): string
     {
         return $this->address['thoroughfare'];
     }
 
-    /**
-     * Get Building Name.
-     *
-     * @return string
-     */
     public function getBuildingName(): string
     {
         return $this->address['building_name'];
     }
 
-    /**
-     * Get Sub Building Name.
-     *
-     * @return string
-     */
     public function getSubBuildingName(): string
     {
         return $this->address['sub_building_name'];
     }
 
-    /**
-     * Get Building Number.
-     *
-     * @return string
-     */
     public function getBuildingNumber(): string
     {
         return $this->address['building_number'];
     }
 
-    /**
-     * Get Sub Building Number.
-     *
-     * @return string
-     */
     public function getSubBuildingNumber(): string
     {
         return $this->address['sub_building_number'];
     }
 
-    /**
-     * Get Line 1.
-     *
-     * @return string
-     */
     public function getLine1(): string
     {
         return $this->address['line_1'];
     }
 
-    /**
-     * Get Line 2.
-     *
-     * @return string
-     */
     public function getLine2(): string
     {
         return $this->address['line_2'];
     }
 
-    /**
-     * Get Line 3.
-     *
-     * @return string
-     */
     public function getLine3(): string
     {
         return $this->address['line_3'];
     }
 
-    /**
-     * Get Line 4.
-     *
-     * @return string
-     */
     public function getLine4(): string
     {
         return $this->address['line_4'];
     }
 
-    /**
-     * Get Line.
-     *
-     * @param int $line
-     *
-     * @return string
-     */
-    public function getLine($line): string
+    public function getLine(int $line): ?string
     {
-        return $this->address['line_'.$line];
+        if ($line > 4) {
+            return null;
+        }
+
+        return $this->address['line_' . $line];
     }
 
-    /**
-     * Get Locality.
-     *
-     * @return string
-     */
     public function getLocality(): string
     {
         return $this->address['locality'];
     }
 
-    /**
-     * Get Town.
-     *
-     * @return string
-     */
     public function getTown(): string
     {
         return $this->address['town_or_city'];
     }
 
-    /**
-     * Get City.
-     *
-     * @return string
-     *
-     * @see ExpandedAddress:getTown()
-     */
     public function getCity(): string
     {
         return $this->address['town_or_city'];
     }
 
-    /**
-     * Get County.
-     *
-     * @return string
-     */
     public function getCounty(): string
     {
         return $this->address['county'];
     }
 
-    /**
-     * Get District.
-     *
-     * @return string
-     */
     public function getDistrict(): string
     {
         return $this->address['district'];
     }
 
-    /**
-     * Get Country.
-     *
-     * @return string
-     */
     public function getCountry(): string
     {
         return $this->address['country'];
@@ -189,51 +96,28 @@ class ExpandedAddress
 
     /**
      * Return a formatted array for the address.
-     *
-     * @param array $keys Override default key names
-     *
-     * @return array
      */
-    public function toArray(array $keys = []): array
+    public function toArray(): array
     {
         return array_merge([
             'formatted_string' => $this->toString(true),
         ], $this->address);
     }
 
-    /**
-     * Returns a string based on the address.
-     *
-     * @param bool $removeEmptyElements Prevents strings having conjoining commas
-     *
-     * @return string
-     */
-    public function toString($removeEmptyElements = false): string
+    public function toString(bool $removeEmptyElements = false): string
     {
-        if (!$removeEmptyElements) {
-            return implode(',', $this->address['formatted_address']);
+        $separator = isset($this->address['formatted_address']) ? ', ' : ',';
+
+        if ($removeEmptyElements === false) {
+            return implode($separator, $this->address['formatted_address']);
         }
 
-        return implode(', ', array_filter($this->address['formatted_address']));
+        return implode($separator, array_filter(
+            $this->address['formatted_address'] ?? $this->address,
+            static fn (?string $value): bool => filled($value)
+        ));
     }
 
-    /**
-     * Compare two addresses to see if they are equal.
-     *
-     * @param \Laralabs\GetAddress\Responses\Address $address Address to compare
-     *
-     * @return bool
-     */
-    public function sameAs(Address $address): bool
-    {
-        return !array_diff($this->address, $address->toArray());
-    }
-
-    /**
-     * Convert the address to a comma separated string.
-     *
-     * @return string
-     */
     public function __toString(): string
     {
         return $this->toString();
